Review param list part 3

+++ include/evmc/evmc.hpp
@@ -320,7 +320,7 @@ constexpr T parse(std::string_view s, std::string_view prefix) noexcept
 // #endif
 
 /// Literal for evmc::address.
-constexpr address operator""_address(const char* s, int) noexcept
+constexpr address operator""_address(const char* s, size_t) noexcept
 {
     return parse<address>(s, "Z");
 }
